crm-fastapi/crm-agent/requirements.txt
fastapi
uvicorn
apscheduler
sqlalchemy[asyncio]
asyncpg
pydantic
python-dotenv
httpx
openai
pytest
pytest-asyncio
pytest-cov

